# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MagicAI  ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/aiimage/versions": {
            "get": {
                "tags": [
                    "AI Image Generation"
                ],
                "summary": "Get Model Versions for AI Image Generation (DALL-E, Stable Diffusion) from settings",
                "description": "Get Model Versions for AI Image Generation",
                "operationId": "versions",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/aiimage/check-availability": {
            "get": {
                "tags": [
                    "AI Image Generation"
                ],
                "summary": "Check if image generation is active / available",
                "description": "Check if image generation is active",
                "operationId": "checkActiveGeneration",
                "responses": {
                    "200": {
                        "description": "Image generation is available."
                    },
                    "409": {
                        "description": "Image generation in progress. Please try again later."
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/aiimage/generate-image": {
            "post": {
                "tags": [
                    "AI Image Generation"
                ],
                "summary": "Generate Image (DALL-E / Stable Diffusion parameters required in request)",
                "description": "Generate Image",
                "operationId": "generateImage",
                "responses": {
                    "200": {
                        "description": "Image generation successful. Image info in json response [images]."
                    },
                    "409": {
                        "description": "Image generation in progress. Please try again later."
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost/oauth/authorize",
                        "tokenUrl": "http://localhost/oauth/token",
                        "refreshUrl": "http://localhost/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    }
}